+++ src/tracker/model/Task.java
@@ -5,7 +5,7 @@ import tracker.controller.InMemoryTaskManager;
 public class Task {
     protected String name;
     protected String description;
-    protected int id;
+    protected long id;
     protected Status status;
 
     public Task() {
@@ -35,11 +35,11 @@ public class Task {
         this.description = description;
     }
 
-    public int getId() {
+    public long getId() {
         return this.id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
